Fixed: language issue, counter for comments, links in views. Added: counter for likes for comment, delete meme. Refactoring: js front-end, back-end
