+++ app/views/homes/_modal-reset-password.haml
@@ -10,9 +10,11 @@
                               "data-bs-dismiss" => "modal",
                               :type => "button"}
           .col-12.d-flex.justify-content-center.mb-4
-            %h3.fw-light Reset password
+            %h3.fw-light
+              = t('pages.index.edit_user.reset_pw')
           .col-12.px-5.mb-4
             %form
               .mb-3
-                %input#resetPasswordEmail.form-control{:placeholder => "Your email", :type => "email"}/
-              %button.btn.btn-primary.w-100.rounded-pill{:type => "submit"} Reset password
+                %input#resetPasswordEmail.form-control{:placeholder => "Email", :type => "email"}/
+              %button.btn.btn-primary.w-100.rounded-pill{:type => "submit"}
+                = t('pages.index.edit_user.reset_pw')
