--- app/views/homes/_navbar-logged-in.haml
@@ -1,55 +1,0 @@
-= render "add-meme-btn-mobile"
-%nav.navbar.navbar-light.bg-light-blue.navbar-expand-lg.fixed-top.hidden-xs.hidden-sm
-  .container-fluid
-    %a#h-logo.navbar-brand{:href => "/"}
-      = t('pages.index.home')
-    %button.navbar-toggler{"aria-controls"  => "offcanvasNavbar",
-                           "data-bs-target" => "#offcanvasNavbar",
-                           "data-bs-toggle" => "offcanvas",
-                           :type => "button"}
-      %span#h-toggler-icon.navbar-toggler-icon
-    #offcanvasNavbar.offcanvas.offcanvas-end.bg-light-blue{"aria-labelledby" => "offcanvasNavbarLabel", :tabindex => "-1"}
-      .offcanvas-header.mb-0.pb-0
-        .user_id_not_security{"user_id" => current_user.id}
-        %h5#offcanvasNavbarLabel.offcanvas-title
-          = t('pages.index.menu')
-        %button.btn-close.text-reset{"aria-label" => "Close",
-                                     "data-bs-dismiss" => "offcanvas",
-                                     :type => "button"}
-      .offcanvas-body
-        %ul.navbar-nav.justify-content-end.flex-grow-1.pe-3
-          %li#lang-selector.nav-item.dropdown.px-0.py-0
-            %a#navbarDropdownMenuLink.nav-link.dropdown-toggle{"aria-expanded"  => "false",
-                                                               "data-bs-toggle" => "dropdown",
-                                                               :href => "#",
-                                                               :role => "button"}
-              - if I18n.locale == :en
-                %img{:src => "/images/lang-usa.png"}/
-              - if I18n.locale == :de
-                %img{:src => "/images/lang-germany.png"}/
-            %ul.dropdown-menu.bg-light-blue{"aria-labelledby" => "navbarDropdownMenuLink"}
-              %li
-                %a.dropdown-item{:href => "/en"}
-                  %img.me-1{:src => "/images/lang-usa.png"}/
-                  = t('pages.index.navbar-lang-en')
-              %li
-                %a.dropdown-item{:href => "/de"}
-                  %img.me-1{:src => "/images/lang-germany.png"}/
-                  = t('pages.index.navbar-lang-de')
-          %li.nav-item
-            = link_to t('menu.addMeme'),
-                        '#',
-                        class: "nav-link my-0 py-0",
-                        "data-bs-target" => "#addMemeModal",
-                        "data-bs-toggle" => "modal"
-          %li.nav-item
-            = link_to t('menu.edit'),
-                        '#',
-                        class: "nav-link my-0 py-0",
-                        "data-bs-target" => "#editUserModal",
-                        "data-bs-toggle" => "modal"
-          %li.nav-item
-            = link_to t('menu.log_out'),
-                      destroy_user_session_path,
-                      method: :delete,
-                      class: "nav-link my-0 py-0"
