+++ config/locales/en.yml
@@ -1,34 +1,3 @@
-# Files in the config/locales directory are used for internationalization
-# and are automatically loaded by Rails. If you want to use locales other
-# than English, add the necessary files in this directory.
-#
-# To use the locales, use `I18n.t`:
-#
-#     I18n.t 'hello'
-#
-# In views, this is aliased to just `t`:
-#
-#     <%= t('hello') %>
-#
-# To use a different locale, set it with `I18n.locale`:
-#
-#     I18n.locale = :es
-#
-# This would use the information in config/locales/es.yml.
-#
-# The following keys must be escaped otherwise they will not be retrieved by
-# the default I18n backend:
-#
-# true, false, on, off, yes, no
-#
-# Instead, surround them with single quotes.
-#
-# en:
-#   'true': 'foo'
-#
-# To learn more, please read the Rails Internationalization guide
-# available at https://guides.rubyonrails.org/i18n.html.
-
 en:
   en: 'English'
   de: 'German'
@@ -50,7 +19,7 @@ en:
       navbar:            'Navigation'
       navbar-lang-en:    'English'
       navbar-lang-de:    'German'
-      home:              'Home'
+      home:              'ME-ME-ME.de'
       menu:              'Menu'
 
       sign_up:
@@ -83,7 +52,7 @@ en:
         rem_acc:     'Do you want to remove your account?'
         rem_acc_btn: 'Delete account'
         update_btn:  'Update'
-        #reset_pw:    'Reset password'
+        reset_pw:    'Reset password'
 
       add_meme:
         title:      'Add a new meme'
@@ -101,6 +70,9 @@ en:
         post_btn:     'Post a comment'
         all_comments: 'All comments for this meme'
         no_comments:  'So far no one has commented on this meme'
+
+      delete_meme:
+        btn-menu: 'Delete meme'
 
       report_for_meme:
         title:    'Report meme'
