+++ config/routes.rb
@@ -1,5 +1,4 @@
 Rails.application.routes.draw do
-  # For details on the DSL available within this file, see https://guides.rubyonrails.org/routing.html
 
   # i18n
   scope '(:locale)', locale: /#{I18n.available_locales.join('|')}/ do
@@ -21,4 +20,5 @@ Rails.application.routes.draw do
       registrations: "custom_registrations"
     }
   end
+
 end
